TWEAK: Unit test refactoring
Split into more files - utils and objects.
Extend NameObject tests with multiple scenarios.
