+++ src/vanillapdf.unittest/document_test.cpp
@@ -1,7 +1,5 @@
 #include "unittest.h"
 
-#include <gtest/gtest.h>
-
 static const unsigned char SIGNING_CERTIFICATE[] = {
     0x30, 0x82, 0x09, 0xc6, 0x02, 0x01, 0x03, 0x30, 0x82, 0x09, 0x90, 0x06, 0x09, 0x2a, 0x86, 0x48,
     0x86, 0xf7, 0x0d, 0x01, 0x07, 0x01, 0xa0, 0x82, 0x09, 0x81, 0x04, 0x82, 0x09, 0x7d, 0x30, 0x82,
